A temporary fix for contract upgrade transactions: during LedgerTransaction verification run the right logic based on whether it contains the UpgradeCommand.
Move ContractUpgradeFlowTest away from createSomeNodes()

Remove assembleBareTx as it's not used
